Split serve() functionality between client and server
